Added more texture functionallity (NPatch), changes to Raymath visibility

+++ src/com/creedvi/raylib/java/rlj/rlgl/VertexBuffer.java
@@ -8,7 +8,7 @@ public class VertexBuffer{
     int cCounter;               // Vertex color counter to process (and draw) from full buffer
     float[] vertices;            // Vertex position (XYZ - 3 components per vertex) (shader-location = 0)
     float[] texcoords;           // Vertex texture coordinates (UV - 2 components per vertex) (shader-location = 1)
-    int[] colors;              // Vertex colors (RGBA - 4 components per vertex) (shader-location = 3)
+    float[] colors;              // Vertex colors (RGBA - 4 components per vertex) (shader-location = 3)
     int[] indices_GL11;      // Vertex indices (in case vertex data comes indexed) (6 indices per quad)
     short[] indices_ES20;    // Vertex indices (in case vertex data comes indexed) (6 indices per quad)
     int vaoId;         // OpenGL Vertex Array Object id
@@ -17,7 +17,7 @@ public class VertexBuffer{
     public VertexBuffer(){
         vertices = new float[3];
         texcoords = new float[2];
-        colors = new int[64];
+        colors = new float[64];
         indices_GL11 = new int[6];
         indices_ES20 = new short[6];
         vboId = new int[4];
@@ -71,11 +71,11 @@ public class VertexBuffer{
         this.texcoords = texcoords;
     }
 
-    public int[] getColors(){
+    public float[] getColors(){
         return colors;
     }
 
-    public void setColors(int[] colors){
+    public void setColors(float[] colors){
         this.colors = colors;
     }
